Move source code into `src` directories and test code inside respective services
