+++ services/conversion/Dockerfile
@@ -9,9 +9,7 @@ RUN pip install --no-cache-dir -r requirements.txt
 RUN ["opentelemetry-bootstrap", "-a", "install"]
 
 # copy the current service source code
-COPY . .
-
-EXPOSE 5000
+COPY src .
 
 # Run the Flask application with opentelemetry instrumentation
 CMD ["opentelemetry-instrument", "python", "main.py"]
