+++ services/conversion/src/conversion/convert.py
@@ -10,7 +10,7 @@ from opentelemetry import metrics, trace
 from opentelemetry.trace import Span
 from opentelemetry.trace.span import StatusCode as SpanStatusCode
 
-from conversion.database.mongo import mongo_collection
+from src.conversion.database.mongo import mongo_collection
 
 tracer = trace.get_tracer(__name__)
 meter = metrics.get_meter(__name__)
