+++ services/conversion/src/main.py
@@ -1,6 +1,6 @@
-
 def _main():
     pass
+
 
 if __name__ == '__main__':
     _main()
