+++ services/conversion/tests/main.py
@@ -1,0 +1,4 @@
+import unittest
+
+if __name__ == "__main__":
+    unittest.main()
