+++ services/conversion/tests/test_connection_mongo.py
@@ -1,7 +1,7 @@
 import unittest
 from unittest.mock import MagicMock
 
-from src.conversion.convert import save_to_mongodb
+from conversion.convert import save_to_mongodb
 
 
 class TestSaveToMongodb(unittest.TestCase):
