+++ services/user_service/src/templates/index.html
