--- tests/main.py
@@ -1,8 +1,0 @@
-import unittest
-from test_connection_mongo import TestSaveToMongodb
-from test_converters import TestTextConversion  # Importing the test class
-
-if __name__ == "__main__":
-    unittest.main()
-
-
